Move ShutdownMonitor into jaxrs
