+++ jaxrs/src/main/java/com/proofpoint/jaxrs/JaxrsModule.java
@@ -56,10 +56,12 @@ import java.util.Set;
 import java.util.concurrent.atomic.AtomicReference;
 
 import static com.google.common.base.Preconditions.checkState;
+import static com.google.inject.Scopes.SINGLETON;
 import static com.google.inject.multibindings.MapBinder.newMapBinder;
 import static com.google.inject.multibindings.Multibinder.newSetBinder;
 import static com.proofpoint.configuration.ConfigurationModule.bindConfig;
 import static com.proofpoint.jaxrs.JaxrsBinder.jaxrsBinder;
+import static com.proofpoint.reporting.HealthBinder.healthBinder;
 
 public class JaxrsModule
         implements Module
@@ -95,6 +97,9 @@ public class JaxrsModule
         jaxrsBinder(binder).bindAdmin(WadlResource.class);
 
         bindConfig(binder).to(JaxrsConfig.class);
+
+        binder.bind(ShutdownMonitor.class).in(SINGLETON);
+        healthBinder(binder).export(ShutdownMonitor.class);
     }
 
     @Provides
